+++ src/iterativeDeepening.cpp
@@ -7,7 +7,7 @@
 namespace IterativeDeepening
 {
 
-    void getMoveIterativeDeepening(uint8_t &x, uint8_t &y, uint8_t **board, uint8_t playerNumber, const std::function<int(uint8_t **, uint8_t)> &heuristic, const std::chrono::time_point<std::chrono::steady_clock> &stopTime){
+    void getMoveIterativeDeepening(uint8_t &x, uint8_t &y, uint8_t **board, uint8_t playerNumber, const std::function<float(uint8_t **, uint8_t)> &heuristic, const std::chrono::time_point<std::chrono::steady_clock> &stopTime){
         uint8_t iterationDepth = 1;
         uint8_t tempx;
         uint8_t tempy;
